requirements.txt
numpy
tensorflow
tensorflow-hub
faiss-cpu
flask
beautifulsoup4
flask_cors